requirements.txt
streamlit
numpy
pillow
tflite-runtime
